@@ -11,13 +11,26 @@ import createReactClass from 'create-react-class';
 import ComponentRegistry from '../src/ComponentRegistry';
 
 describe('ComponentRegistry', () => {
-  expect.assertions(9);
-  it('registers and retrieves generator function components', () => {
+  beforeEach(() => {
+    ComponentRegistry.components().clear();
+  });
+
+  it('registers and retrieves React function components', () => {
     expect.assertions(1);
     const C1 = () => <div>HELLO</div>;
     ComponentRegistry.register({ C1 });
     const actual = ComponentRegistry.get('C1');
-    const expected = { name: 'C1', component: C1, generatorFunction: true, isRenderer: false };
+    const expected = { name: 'C1', component: C1, renderFunction: false, isRenderer: false };
+    expect(actual).toEqual(expected);
+  });
+
+  it('registers and retrieves render function components where property renderFunction is set and zero params', () => {
+    expect.assertions(1);
+    const C1 = () => <div>HELLO</div>;
+    C1.renderFunction = true;
+    ComponentRegistry.register({ C1 });
+    const actual = ComponentRegistry.get('C1');
+    const expected = { name: 'C1', component: C1, renderFunction: true, isRenderer: false };
     expect(actual).toEqual(expected);
   });
 
@@ -30,7 +43,7 @@ describe('ComponentRegistry', () => {
     });
     ComponentRegistry.register({ C2 });
     const actual = ComponentRegistry.get('C2');
-    const expected = { name: 'C2', component: C2, generatorFunction: false, isRenderer: false };
+    const expected = { name: 'C2', component: C2, renderFunction: false, isRenderer: false };
     expect(actual).toEqual(expected);
   });
 
@@ -43,16 +56,16 @@ describe('ComponentRegistry', () => {
     }
     ComponentRegistry.register({ C3 });
     const actual = ComponentRegistry.get('C3');
-    const expected = { name: 'C3', component: C3, generatorFunction: false, isRenderer: false };
+    const expected = { name: 'C3', component: C3, renderFunction: false, isRenderer: false };
     expect(actual).toEqual(expected);
   });
 
-  it('registers and retrieves renderers', () => {
+  it('registers and retrieves renderers if 3 params', () => {
     expect.assertions(1);
     const C4 = (a1, a2, a3) => null;
     ComponentRegistry.register({ C4 });
     const actual = ComponentRegistry.get('C4');
-    const expected = { name: 'C4', component: C4, generatorFunction: true, isRenderer: true };
+    const expected = { name: 'C4', component: C4, renderFunction: true, isRenderer: true };
     expect(actual).toEqual(expected);
   });
 
@@ -61,23 +74,33 @@ describe('ComponentRegistry', () => {
    * Thus, tests are cummulative.
    */
   it('registers and retrieves multiple components', () => {
-    expect.assertions(3);
+    expect.assertions(4);
+    // Plain react stateless functional components
     const C5 = () => <div>WHY</div>;
     const C6 = () => <div>NOW</div>;
+    const C7 = () => <div>NOW</div>;
+    C7.renderFunction = true;
     ComponentRegistry.register({ C5 });
     ComponentRegistry.register({ C6 });
+    ComponentRegistry.register({ C7 });
     const components = ComponentRegistry.components();
-    expect(components.size).toBe(6);
+    expect(components.size).toBe(3);
     expect(components.get('C5')).toEqual({
       name: 'C5',
       component: C5,
-      generatorFunction: true,
+      renderFunction: false,
       isRenderer: false,
     });
     expect(components.get('C6')).toEqual({
       name: 'C6',
       component: C6,
-      generatorFunction: true,
+      renderFunction: false,
+      isRenderer: false,
+    });
+    expect(components.get('C7')).toEqual({
+      name: 'C7',
+      component: C7,
+      renderFunction: true,
       isRenderer: false,
     });
   });
@@ -92,13 +115,13 @@ describe('ComponentRegistry', () => {
     expect(components.get('C7')).toEqual({
       name: 'C7',
       component: C7,
-      generatorFunction: true,
+      renderFunction: true,
       isRenderer: false,
     });
     expect(components.get('C8')).toEqual({
       name: 'C8',
       component: C8,
-      generatorFunction: true,
+      renderFunction: false,
       isRenderer: false,
     });
   });
