@@ -4,14 +4,17 @@
 import React from 'react';
 
 import serverRenderReactComponent from '../src/serverRenderReactComponent';
-import ComponentStore from '../src/ComponentRegistry';
+import ComponentRegistry from '../src/ComponentRegistry';
 
 describe('serverRenderReactComponent', () => {
-  expect.assertions(7);
+  beforeEach(() => {
+    ComponentRegistry.components().clear();
+  });
+
   it('serverRenderReactComponent renders a registered component', () => {
     expect.assertions(2);
     const X1 = () => <div>HELLO</div>;
-    ComponentStore.register({ X1 });
+    ComponentRegistry.register({ X1 });
 
     const { html, hasErrors } = JSON.parse(
       serverRenderReactComponent({ name: 'X1', domNodeId: 'myDomId', trace: false }),
@@ -28,7 +31,7 @@ describe('serverRenderReactComponent', () => {
       throw new Error('XYZ');
     };
 
-    ComponentStore.register({ X2 });
+    ComponentRegistry.register({ X2 });
 
     // Not testing the consoleReplayScript, as handleError is putting the console to the test
     // runner log.
@@ -44,9 +47,9 @@ describe('serverRenderReactComponent', () => {
   it('serverRenderReactComponent renders html', () => {
     expect.assertions(2);
     const expectedHtml = '<div>Hello</div>';
-    const X3 = () => ({ renderedHtml: expectedHtml });
+    const X3 = (props, _railsContext) => ({ renderedHtml: expectedHtml });
 
-    ComponentStore.register({ X3 });
+    ComponentRegistry.register({ X3 });
 
     const { html, hasErrors, renderedHtml } = JSON.parse(
       serverRenderReactComponent({ name: 'X3', domNodeId: 'myDomId', trace: false }),
@@ -59,7 +62,7 @@ describe('serverRenderReactComponent', () => {
   it('serverRenderReactComponent renders an error if attempting to render a renderer', () => {
     expect.assertions(1);
     const X3 = (a1, a2, a3) => null;
-    ComponentStore.register({ X3 });
+    ComponentRegistry.register({ X3 });
 
     const { html } = JSON.parse(
       serverRenderReactComponent({ name: 'X3', domNodeId: 'myDomId', trace: false }),
