@@ -9,7 +9,14 @@ import EchoProps from '../components/EchoProps';
  *  React will see that the state is the same and not do anything.
  *  Note, this is imported as "HelloWorldApp" by "clientRegistration.jsx"
  *
- *  Note, this is a fictional example, as you'd only use a generator function if you wanted to run
+ *  Note, this is a fictional example, as you'd only use a render function if you wanted to run
  *  some extra code, such as setting up Redux and React-Router.
  */
-export default (props, _railsContext) => <EchoProps {...props} />;
+
+// This will not work:
+// export default (props, _railsContext) => () => <EchoProps {...props} />;
+
+// You may do either:
+// export default (props, _railsContext) => () => <EchoProps {...props} />;
+// or
+export default (props) => <EchoProps {...props} />;
