@@ -2,4 +2,12 @@
 import React from 'react';
 import ReactHelmet from '../components/ReactHelmet';
 
-export default (props, _railsContext) => <ReactHelmet {...props} />;
+// This works fine, React functional component:
+// export default (props) => <ReactHelmet {...props} />;
+
+export default (props) => <ReactHelmet {...props} />;
+
+// Note, the server side has to be a render function
+
+// If you want a renderFunction, return a ReactComponent
+// export default (props, _railsContext) => () => <ReactHelmet {...props} />;
