@@ -1,4 +1,6 @@
 // Top level component for simple client side only rendering
+// This one is broken in that the function takes one param, so it's not a generator
+// function. The point of this is to provide a good error.
 import React from 'react';
 import { renderToString } from 'react-dom/server';
 import { Helmet } from 'react-helmet';
@@ -12,10 +14,10 @@ import ReactHelmet from '../components/ReactHelmet';
  *  This is imported as "ReactHelmetApp" by "serverRegistration.jsx". Note that rendered
  *  component markup must go under "componentHtml" key.
  *
- *  Note that the function takes 2 params to identify this as a generator fuction. Alternately,
- *  the function could get the property of `.renderFunction = true` added to it.
+ *  Note that the function should take 2 params to identify this as a generator fuction.
+ *  Alternately, the function could get the property of `.renderFunction = true` added to it.
  */
-export default (props, _railsContext) => {
+export default (props) => {
   const componentHtml = renderToString(<ReactHelmet {...props} />);
   const helmet = Helmet.renderStatic();
 
@@ -23,7 +25,5 @@ export default (props, _railsContext) => {
     componentHtml,
     title: helmet.title.toString(),
   };
-
-  // Note that this function returns an Object for server rendering.
   return { renderedHtml };
 };
