@@ -6,12 +6,14 @@ import HelloString from '../non_react/HelloString';
 
 // React components
 import HelloWorld from '../components/HelloWorld';
+import HelloWorldHooks from '../components/HelloWorldHooks';
+import HelloWorldHooksContext from '../components/HelloWorldHooksContext';
 
 import HelloWorldES5 from '../components/HelloWorldES5';
 import HelloWorldRehydratable from '../components/HelloWorldRehydratable';
 import HelloWorldWithLogAndThrow from '../components/HelloWorldWithLogAndThrow';
 
-// Generator function
+// render function
 import HelloWorldApp from '../startup/HelloWorldApp';
 import BrokenApp from '../startup/BrokenApp';
 
@@ -33,11 +35,14 @@ import SharedReduxStore from '../stores/SharedReduxStore';
 import DeferredRenderApp from '../startup/DeferredRenderAppServer';
 import ManualRenderApp from '../startup/ManualRenderAppRenderer';
 
-// Deferred render on the client side w/ server render
 import RenderedHtml from '../startup/ServerRenderedHtml';
 
-// Deferred render on the client side w/ server render with additional HTML strings:
+// Server render with additional HTML strings:
 import ReactHelmetApp from '../startup/ReactHelmetServerApp';
+
+// Broken server render since ReactHelmetServerAppBroken is not properly defined
+// to be a renderFunction
+import ReactHelmetAppBroken from '../startup/ReactHelmetServerAppBroken';
 
 // Demonstrate using Images
 import ImageExample from '../components/ImageExample';
@@ -64,8 +69,11 @@ ReactOnRails.register({
   DeferredRenderApp,
   RenderedHtml,
   ReactHelmetApp,
+  ReactHelmetAppBroken,
   ImageExample,
   SetTimeoutLoggingApp,
+  HelloWorldHooks,
+  HelloWorldHooksContext,
 });
 
 ReactOnRails.registerStore({
